hotel-booking-api/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "hotel_booking_api"
version = "0.1.0"
description = "Hotel Booking API with MCP server support"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "email-validator",
  "python-dotenv",
  "mcp",
  "pyjwt",
]

[project.scripts]
hotel-api = "hotel_booking_api.main:main"

[dependency-groups]
dev = [
  "pytest",
  "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["hotel_booking_api"]

[tool.hatch.build.targets.wheel.force-include]
"hotel_booking_api/accommodations.json" = "hotel_booking_api/accommodations.json"
"hotel_booking_api/bookings.json" = "hotel_booking_api/bookings.json"
